[Refactor] createElem util 생성
