+++ 02.typescript/todoApp/src/pages/regist/TodoRegist.ts
@@ -12,26 +12,6 @@ const TodoRegist = function () {
   const content = document.createElement("div");
   content.setAttribute("id", "content");
 
-  const createElem = (
-    parent: HTMLElement,
-    tagName: string,
-    txt: string = "",
-    ...attributes: [string, string][]
-  ) => {
-    const element = document.createElement(tagName);
-
-    attributes.forEach(([attrName, attrValue]) => {
-      element.setAttribute(attrName, attrValue);
-    });
-
-    if (txt) {
-      element.textContent = txt;
-    }
-
-    parent.appendChild(element);
-    return element; // 생성된 요소 반환
-  };
-
   const form = createElem(content, "form", "", ["id", "form"]);
 
   const fieldset = createElem(form, "fieldset", "", ["id", "fieldset"])
